# docs/swagger.yaml
definitions:
...
  model.HTTPMessage:
    description: HTTPMessage represents a message to be sent to the client trough
      HTTP REST API.
    properties:
      message:
        example: success
        format: string
        type: string
      method:
        example: GET
        format: string
        type: string
      path:
        example: /api/v1/users
        format: string
        type: string
      status_code:
        example: 200
        format: int32
        type: integer
      timestamp:
        example: "2021-07-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
  model.ListResourcesResponse:
    description: ListResourcesResponse represents a list of users.
    properties:
      items:
        items:
          $ref: '#/definitions/model.Resource'
        type: array
      paginator:
        $ref: '#/definitions/model.Paginator'
    type: object
...
  model.Paginator:
    description: Paginator represents a paginator.
    properties:
      limit:
        example: 10
        format: int
        type: integer
      next_page:
        example: http://localhost:8080/users?next_token=ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=&limit=10
        format: string
        type: string
      next_token:
        example: ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=
        format: string
        type: string
      prev_page:
        example: http://localhost:8080/users?prev_token=ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=&limit=10
        format: string
        type: string
      prev_token:
        example: ZmZmZmZmZmYtZmZmZi0tZmZmZmZmZmY=
        format: string
        type: string
      size:
        example: 10
        format: int
        type: integer
    type: object
...
  model.Resource:
    description: Resource represents a permission.
    properties:
      action:
        example: GET
        format: string
        type: string
      created_at:
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
      description:
        example: Allows reading of users
        format: string
        type: string
      id:
        example: 01979cde-6d91-7743-a481-bfe67f26a3c2
        format: uuid
        type: string
      name:
        example: Read Users
        format: string
        type: string
      resource:
        example: users
        format: string
        type: string
      system:
        example: false
        format: bool
        type: boolean
      updated_at:
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
info:
  contact:
    email: info@qu3ry.me
    name: API Support
    url: https://qu3ry.me
  description: |-
    This is a service template for building RESTful APIs in Go.
    It uses a PostgreSQL database to store user information.
    The service provides:
    - CRUD operations for users.
    - Health and version endpoints.
    - Configuration using environment variables or command line arguments.
    - Debug mode to enable debug logging.
    - TLS enabled to secure the communication.
  title: Go REST API Service Template
  version: v1
paths:
...
  /resources:
    get:
      description: Retrieve paginated list of all resources in the system
      operationId: 019791cc-06c7-7e8e-8d7e-cd3f9296e0fd
      parameters:
      - description: 'Comma-separated list of fields to sort by. Example: first_name
          ASC, created_at DESC'
        format: string
        in: query
        name: sort
        type: string
      - description: 'Filter field. Example: id=1 AND first_name=''John'''
        format: string
        in: query
        name: filter
        type: string
      - description: 'Fields to return. Example: id,first_name,last_name'
        format: string
        in: query
        name: fields
        type: string
      - description: Next cursor
        format: string
        in: query
        name: next_token
        type: string
      - description: Previous cursor
        format: string
        in: query
        name: prev_token
        type: string
      - description: Limit
        format: int
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListResourcesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPMessage'
      security:
      - AccessToken: []
      summary: List resources
      tags:
      - Resources
  /resources/{resource_id}:
    get:
      description: Retrieve a specific resource by its identifier
      operationId: 019791cc-06c7-7e86-ad42-b777bfcc9e40
      parameters:
      - description: The permission id in UUID format
        format: uuid
        in: path
        name: resource_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Resource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPMessage'
      security:
      - AccessToken: []
      summary: Get resource
      tags:
      - Resources
  /resources/matches:
    get:
      description: Find resources that match specific action and resource policy patterns
      operationId: 019791cc-06c7-7e92-9152-cb35902f79c4
      parameters:
      - description: Action to filter by
        format: string
        in: query
        name: action
        required: true
        type: string
      - description: Resource to filter by
        format: string
        in: query
        name: resource
        required: true
        type: string
      - description: 'Comma-separated list of fields to sort by. Example: first_name
          ASC, created_at DESC'
        format: string
        in: query
        name: sort
        type: string
      - description: 'Fields to return. Example: id,first_name,last_name'
        format: string
        in: query
        name: fields
        type: string
      - description: Next cursor
        format: string
        in: query
        name: next_token
        type: string
      - description: Previous cursor
        format: string
        in: query
        name: prev_token
        type: string
      - description: Limit
        format: int
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListResourcesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HTTPMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPMessage'
      security:
      - AccessToken: []
      summary: Match resources
      tags:
      - Resources
...
swagger: "2.0"